+++ CHANGELOG.md
@@ -13,6 +13,8 @@ Given a version number MAJOR.MINOR.PATCH, increment:
 
 
 ## [Unreleased]
+### Added
+- new fields to PixChargeback and PixInfraction resources
 
 ## [0.10.0] - 2025-07-23
 ### Added